+++ src/finite_volume/finite_volume/finite_volume.h
@@ -22,7 +22,7 @@ public:
 
     FiniteVolume(const GridBlock<T>& grid, json config);
 
-    int compute_dudt(FlowStates<T>& flow_state, const GridBlock<T>& grid,
+    size_t compute_dudt(FlowStates<T>& flow_state, const GridBlock<T>& grid,
                      ConservedQuantities<T>& dudt, IdealGas<T>& gas_model);
 
     double estimate_dt(const FlowStates<T>& flow_state, GridBlock<T>& grid,
@@ -34,7 +34,7 @@ public:
     void apply_pre_reconstruction_bc(FlowStates<T>& fs,
                                      const GridBlock<T>& grid);
     void reconstruct(FlowStates<T>& flow_states, const GridBlock<T>& grid,
-                     IdealGas<T>& gas_model, unsigned int order);
+                     IdealGas<T>& gas_model, size_t order);
     void copy_reconstruct(FlowStates<T>& flow_states, const GridBlock<T>& grid);
     void linear_reconstruct(FlowStates<T>& flow_states,
                             const GridBlock<T>& grid, IdealGas<T>& gas_model);
@@ -43,7 +43,7 @@ public:
     void compute_flux(const GridBlock<T>& grid, IdealGas<T>& gas_model);
     void apply_post_convective_flux_bc();
     void apply_pre_spatial_deriv();
-    int count_bad_cells(const FlowStates<T>& fs, const int num_cells);
+    size_t count_bad_cells(const FlowStates<T>& fs, const size_t num_cells);
 
 private:
     // memory
@@ -53,11 +53,11 @@ private:
 
     // boundary conditions
     std::vector<std::shared_ptr<BoundaryCondition<T>>> bcs_{};
-    std::vector<Field<int>> bc_interfaces_{};
+    std::vector<Field<size_t>> bc_interfaces_{};
 
     // configuration
-    unsigned int dim_;
-    unsigned int reconstruction_order_;
+    size_t dim_;
+    size_t reconstruction_order_;
     FluxCalculator flux_calculator_;
 
     // gradients
