+++ src/util/util/vector3.cpp
@@ -13,7 +13,7 @@ void dot(const Vector3s<T> &a, const Vector3s<T> &b, Field<T> &result) {
     assert((a.size() == b.size()) && (b.size() == result.size()));
 
     Kokkos::parallel_for(
-        "vector3 dot product", a.size(), KOKKOS_LAMBDA(const int i) {
+        "vector3 dot product", a.size(), KOKKOS_LAMBDA(const size_t i) {
             result(i) =
                 a(i, 0) * b(i, 0) + a(i, 1) * b(i, 1) + a(i, 2) * b(i, 2);
         });
@@ -26,7 +26,7 @@ void add(const Vector3s<T> &a, const Vector3s<T> &b, Vector3s<T> &result) {
     assert((a.size() == b.size()) && (b.size() == result.size()));
 
     Kokkos::parallel_for(
-        "Vector3s add", a.size(), KOKKOS_LAMBDA(const int i) {
+        "Vector3s add", a.size(), KOKKOS_LAMBDA(const size_t i) {
             result(i, 0) = a(i, 0) + b(i, 0);
             result(i, 1) = a(i, 1) + b(i, 1);
             result(i, 2) = a(i, 2) + b(i, 2);
@@ -40,7 +40,7 @@ void subtract(const Vector3s<T> &a, const Vector3s<T> &b, Vector3s<T> &result) {
     assert((a.size() == b.size()) && (b.size() == result.size()));
 
     Kokkos::parallel_for(
-        "Vector3s subtract", a.size(), KOKKOS_LAMBDA(const int i) {
+        "Vector3s subtract", a.size(), KOKKOS_LAMBDA(const size_t i) {
             result(i, 0) = a(i, 0) - b(i, 0);
             result(i, 1) = a(i, 1) - b(i, 1);
             result(i, 2) = a(i, 2) - b(i, 2);
@@ -55,7 +55,7 @@ void cross(const Vector3s<T> &a, const Vector3s<T> &b, Vector3s<T> &result) {
     assert((a.size() == b.size()) && (b.size() == result.size()));
 
     Kokkos::parallel_for(
-        "Vector3s cross", a.size(), KOKKOS_LAMBDA(const int i) {
+        "Vector3s cross", a.size(), KOKKOS_LAMBDA(const size_t i) {
             result(i, 0) = a(i, 1) * b(i, 2) - a(i, 2) * b(i, 1);
             result(i, 1) = a(i, 2) * b(i, 0) - a(i, 0) * b(i, 2);
             result(i, 2) = a(i, 0) * b(i, 1) - a(i, 1) * b(i, 0);
@@ -68,7 +68,7 @@ template void cross<double>(const Vector3s<double> &a,
 template <typename T>
 void scale_in_place(Vector3s<T> &a, T factor) {
     Kokkos::parallel_for(
-        "Vector3s scale", a.size(), KOKKOS_LAMBDA(const int i) {
+        "Vector3s scale", a.size(), KOKKOS_LAMBDA(const size_t i) {
             a(i, 0) *= factor;
             a(i, 1) *= factor;
             a(i, 2) *= factor;
@@ -81,7 +81,7 @@ void length(const Vector3s<T> &a, Field<T> &len) {
     assert(a.size() == len.size());
 
     Kokkos::parallel_for(
-        "Vector3s length", a.size(), KOKKOS_LAMBDA(const int i) {
+        "Vector3s length", a.size(), KOKKOS_LAMBDA(const size_t i) {
             len(i) =
                 sqrt(a(i, 0) * a(i, 0) + a(i, 1) * a(i, 1) + a(i, 2) * a(i, 2));
         });
@@ -91,7 +91,7 @@ template void length<double>(const Vector3s<double> &a, Field<double> &len);
 template <typename T>
 void normalise(Vector3s<T> &a) {
     Kokkos::parallel_for(
-        "Vector3s normalise", a.size(), KOKKOS_LAMBDA(const int i) {
+        "Vector3s normalise", a.size(), KOKKOS_LAMBDA(const size_t i) {
             double length_inv =
                 1. /
                 sqrt(a(i, 0) * a(i, 0) + a(i, 1) * a(i, 1) + a(i, 2) * a(i, 2));
@@ -107,7 +107,7 @@ void transform_to_local_frame(Vector3s<T> &a, const Vector3s<T> &norm,
                               const Vector3s<T> &tan1, const Vector3s<T> tan2) {
     Kokkos::parallel_for(
         "Vector3s::transform_to_local_frame", a.size(),
-        KOKKOS_LAMBDA(const int i) {
+        KOKKOS_LAMBDA(const size_t i) {
             T x = a.x(i) * norm.x(i) + a.y(i) * norm.y(i) + a.z(i) * norm.z(i);
             T y = a.x(i) * tan1.x(i) + a.y(i) * tan1.y(i) + a.z(i) * tan1.z(i);
             T z = a.x(i) * tan2.x(i) + a.y(i) * tan2.y(i) + a.z(i) * tan2.z(i);
@@ -127,7 +127,7 @@ void transform_to_global_frame(Vector3s<T> &a, const Vector3s<T> &norm,
                                const Vector3s<T> &tan2) {
     Kokkos::parallel_for(
         "Vector3s::transform_to_global_frame", a.size(),
-        KOKKOS_LAMBDA(const int i) {
+        KOKKOS_LAMBDA(const size_t i) {
             T x = a(i, 0) * norm(i, 0) + a(i, 1) * tan1(i, 0) +
                   a(i, 2) * tan2(i, 0);
             T y = a(i, 0) * norm(i, 1) + a(i, 1) * tan1(i, 1) +
@@ -146,7 +146,7 @@ template void transform_to_global_frame<double>(Vector3s<double> &a,
                                                 const Vector3s<double> &tan2);
 
 TEST_CASE("Vector Dot Product") {
-    int n = 10;
+    size_t n = 10;
 
     // device memory
     Vector3s<double> a_dev("a", n);
@@ -160,7 +160,7 @@ TEST_CASE("Vector Dot Product") {
     Field<double>::mirror_type expected("expected", n);
 
     // set some data
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         a_host(i, 0) = 1.0 * i;
         a_host(i, 1) = 2.0 * i;
         a_host(i, 2) = 3.0 * i;
@@ -185,13 +185,13 @@ TEST_CASE("Vector Dot Product") {
     result_host.deep_copy(result_dev);
 
     // check the results are correct
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         CHECK(Kokkos::fabs(expected(i) < result_host(i)) < VEC3_TOL);
     }
 }
 
 TEST_CASE("Vector3s Add") {
-    int n = 20;
+    size_t n = 20;
 
     // allocate device memory
     Vector3s<double> a_dev("a", n);
@@ -205,7 +205,7 @@ TEST_CASE("Vector3s Add") {
     Vector3s<double>::mirror_type expected("expected", n);
 
     // set some data
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         a_host(i, 0) = 1.0 * i;
         a_host(i, 1) = -2.0 * i;
         a_host(i, 2) = 3.0 * i - 5;
@@ -230,7 +230,7 @@ TEST_CASE("Vector3s Add") {
     result_host.deep_copy(result_dev);
 
     // check the results are correct
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         CHECK(Kokkos::fabs(expected(i, 0) - result_host(i, 0)) < VEC3_TOL);
         CHECK(Kokkos::fabs(expected(i, 1) - result_host(i, 1)) < VEC3_TOL);
         CHECK(Kokkos::fabs(expected(i, 2) - result_host(i, 2)) < VEC3_TOL);
@@ -238,7 +238,7 @@ TEST_CASE("Vector3s Add") {
 }
 
 TEST_CASE("Vector3s subtract") {
-    int n = 20;
+    size_t n = 20;
     // allocate memory on the device
     Vector3s<double> a_dev("a", n);
     Vector3s<double> b_dev("b", n);
@@ -251,7 +251,7 @@ TEST_CASE("Vector3s subtract") {
     Vector3s<double>::mirror_type expected("expected", n);
 
     // set some data
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         a_host(i, 0) = 1.0 * i;
         a_host(i, 1) = -2.0 * i;
         a_host(i, 2) = 3.0 * i - 5;
@@ -276,7 +276,7 @@ TEST_CASE("Vector3s subtract") {
     result_host.deep_copy(result_dev);
 
     // check the results are correct
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         CHECK(Kokkos::fabs(expected(i, 0) - result_host(i, 0)) < VEC3_TOL);
         CHECK(Kokkos::fabs(expected(i, 1) - result_host(i, 1)) < VEC3_TOL);
         CHECK(Kokkos::fabs(expected(i, 2) - result_host(i, 2)) < VEC3_TOL);
@@ -284,7 +284,7 @@ TEST_CASE("Vector3s subtract") {
 }
 
 TEST_CASE("Vector3s cross") {
-    int n = 20;
+    size_t n = 20;
 
     // allocate memory on the device
     Vector3s<double> a_dev("a", n);
@@ -298,7 +298,7 @@ TEST_CASE("Vector3s cross") {
     Vector3s<double>::mirror_type expected("expected", n);
 
     // set some data
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         a_host(i, 0) = 1.0 * i;
         a_host(i, 1) = -2.0 * i;
         a_host(i, 2) = 3.0 * i - 5;
@@ -323,7 +323,7 @@ TEST_CASE("Vector3s cross") {
     result_host.deep_copy(result_dev);
 
     // check the results are correct
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         CHECK(Kokkos::fabs(expected(i, 0) - result_host(i, 0)) < VEC3_TOL);
         CHECK(Kokkos::fabs(expected(i, 1) - result_host(i, 1)) < VEC3_TOL);
         CHECK(Kokkos::fabs(expected(i, 2) - result_host(i, 2)) < VEC3_TOL);
@@ -331,12 +331,12 @@ TEST_CASE("Vector3s cross") {
 }
 
 TEST_CASE("Vector3s scale_in_place") {
-    int n = 20;
+    size_t n = 20;
     Vector3s<double> a_dev("a", n);
     auto a_host = a_dev.host_mirror();
     double factor = 2.0;
 
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         a_host(i, 0) = 1.0 * i;
         a_host(i, 1) = 2.0 * i;
         a_host(i, 2) = 3.0 * i;
@@ -346,7 +346,7 @@ TEST_CASE("Vector3s scale_in_place") {
     scale_in_place(a_dev, factor);
     a_host.deep_copy(a_dev);
 
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         CHECK(Kokkos::fabs(a_host.x(i) - 2.0 * i) < VEC3_TOL);
         CHECK(Kokkos::fabs(a_host.y(i) - 4.0 * i) < VEC3_TOL);
         CHECK(Kokkos::fabs(a_host.z(i) - 6.0 * i) < VEC3_TOL);
@@ -354,7 +354,7 @@ TEST_CASE("Vector3s scale_in_place") {
 }
 
 TEST_CASE("Vector3s length") {
-    int n = 20;
+    size_t n = 20;
     Vector3s<double> a_dev("a", n);
     Field<double> len_dev("length", n);
 
@@ -362,7 +362,7 @@ TEST_CASE("Vector3s length") {
     auto len_host = len_dev.host_mirror();
     Field<double>::mirror_type expected("expected", n);
 
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         a_host(i, 0) = 1.0 * i;
         a_host(i, 1) = 2.0 * i;
         a_host(i, 2) = 3.0 * i;
@@ -376,17 +376,17 @@ TEST_CASE("Vector3s length") {
     length(a_dev, len_dev);
     len_host.deep_copy(len_dev);
 
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         CHECK(Kokkos::fabs(len_host(i) - expected(i)) < VEC3_TOL);
     }
 }
 
 TEST_CASE("Vector3s normalise") {
-    int n = 20;
+    size_t n = 20;
     Vector3s<double> a_dev("a", n);
     auto a_host = a_dev.host_mirror();
 
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         a_host.x(i) = 1.0 * i + 1.0;
         a_host.y(i) = 2.0 * i;
         a_host.z(i) = 3.0 * i;
@@ -396,7 +396,7 @@ TEST_CASE("Vector3s normalise") {
     normalise(a_dev);
     a_host.deep_copy(a_dev);
 
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         double length_inv =
             sqrt((i + 1.0) * (i + 1.0) + 4.0 * i * i + 9.0 * i * i);
         CHECK(Kokkos::fabs(a_host.x(i) - 1.0 * (i + 1.0) / length_inv) <
@@ -407,7 +407,7 @@ TEST_CASE("Vector3s normalise") {
 }
 
 TEST_CASE("Vector3s::transform_to_local_frame") {
-    int n = 3;
+    size_t n = 3;
     Vector3s<double> a_dev("a", n);
     Vector3s<double> norm_dev("norm", n);
     Vector3s<double> tan1_dev("tan1", n);
